merge

+++ modules/PSBuild/PSBuild.psm1
@@ -1565,7 +1565,7 @@ function Build-PSScript
 									#Check if NugetPackage is already downloaded
 									try
 									{
-										$ModDependencyExcepctedPath = Join-Path -Path $DestinationPath -ChildPath $dependantModuleName -ErrorAction Stop
+										$ModDependencyExcepctedPath = Join-Path -Path $DependencyDestinationPath -ChildPath $dependantModuleName -ErrorAction Stop
 										Remove-Variable -Name ModDependencyExist -ErrorAction SilentlyContinue
 										$ModDependencyExist = Get-Module -ListAvailable -FullyQualifiedName $ModDependencyExcepctedPath -Refresh -ErrorAction Stop -Verbose:$false
 									}
@@ -1585,15 +1585,15 @@ function Build-PSScript
 
 										#Downloading NugetPackage
 										Write-Verbose "Build Script:$scriptName/$scriptVersion in progress. Build dependant module:$dependantModuleName/$dependantModuleVersion in progress. Downloading PSGetPackage: $($NuGetDependancyHandle.Name)/$($NuGetDependancyHandle.Version)"
-										if (-not (Test-Path $DestinationPath))
+										if (-not (Test-Path $DependencyDestinationPath))
 										{
-											$null = New-Item -Path $DestinationPath -ItemType Directory -ErrorAction Stop
+											$null = New-Item -Path $DependencyDestinationPath -ItemType Directory -ErrorAction Stop
 										}
 										$PSGet_Params = @{
 											Name=$dependantModuleName
 											Repository=$ModuleRepo.Name
 											RequiredVersion=$NuGetDependancyHandle.Version
-                                            Path=$DestinationPath
+                                            Path=$DependencyDestinationPath
 										}
 										if ($ModuleRepo.ContainsKey('Credential'))
 										{
@@ -1882,19 +1882,21 @@ function Build-PSSolution
 		{
 			Write-Verbose "Configure PS Environment started"
 
-			$TempModulePaths = $SolutionConfig.SolutionStructure.ModulesPath.BuildPath
-			if ($SolutionConfig.Build.AutoloadbuiltModulesForUser)
+			if ($SolutionConfig.SolutionStructure.ModulesPath.BuildPath)
 			{
+				$TempModulePaths = $SolutionConfig.SolutionStructure.ModulesPath.BuildPath
+				if ($SolutionConfig.Build.AutoloadbuiltModulesForUser)
+				{
 
-				Write-Verbose "Configure PS Environment in progress. Enable BuildPath($($TempModulePaths -join ',')) Module Autoloading"
-				Add-PSModulePathEntry -Path $TempModulePaths -Scope User,Process -ErrorAction Stop -Force
+					Write-Verbose "Configure PS Environment in progress. Enable BuildPath($($TempModulePaths -join ',')) Module Autoloading"
+					Add-PSModulePathEntry -Path $TempModulePaths -Scope User,Process -ErrorAction Stop -Force
+				}
+				else
+				{
+					Write-Verbose "Configure PS Environment in progress. Disable BuildPath($($TempModulePaths -join ',')) Module Autoloading"
+					Remove-PSModulePathEntry -Path $TempModulePaths -Scope User,Process -ErrorAction Stop -WarningAction SilentlyContinue
+				}
 			}
-			else
-			{
-				Write-Verbose "Configure PS Environment in progress. Disable BuildPath($($TempModulePaths -join ',')) Module Autoloading"
-				Remove-PSModulePathEntry -Path $TempModulePaths -Scope User,Process -ErrorAction Stop -WarningAction SilentlyContinue
-			}
-
 			Write-Verbose "Configure PS Environment completed"
 		}
 		catch
@@ -1905,37 +1907,44 @@ function Build-PSSolution
 		#Build Solution Modules
 		try
 		{
-			Write-Verbose "Build PSModules started"
-
 			[ref]$ModuleValidationCache = @{}
 			[ref]$PsGetModuleValidationCache = @{}
 
-			#Validate All Modules
-			$AllModulesPath = New-Object System.Collections.ArrayList -ErrorAction Stop
-			foreach ($ModulePath in $SolutionConfig.SolutionStructure.ModulesPath)
+			if ($SolutionConfig.SolutionStructure.ModulesPath)
 			{
-				$null = $AllModulesPath.AddRange((Get-ChildItem -Path $ModulePath.SourcePath -Directory -ErrorAction Stop))
-			}
-			priv_Validate-Module -SourcePath $AllModulesPath -ModuleValidationCache $ModuleValidationCache
+				Write-Verbose "Build PSModules started"
 
-			foreach ($ModulePath in $SolutionConfig.SolutionStructure.ModulesPath)
-			{
-				$Modules = Get-ChildItem -Path $ModulePath.SourcePath -Directory -ErrorAction Stop
-				$BuildPSSolutionModule_Params = @{
-					SourcePath=$Modules
-					DestinationPath=$ModulePath.BuildPath
-					ResolveDependancies=$SolutionConfig.Build.AutoResolveDependantModules
-					PSGetRepository=$SolutionConfig.Packaging.PSGetSearchRepositories
-					CheckCommandReferences=$SolutionConfig.Build.CheckCommandReferences.Enabled
-					CheckCommandReferencesConfiguration=$SolutionConfig.Build.CheckCommandReferences
-					ModuleValidationCache=$ModuleValidationCache
-					UpdateModuleReferences=$SolutionConfig.Build.UpdateModuleReferences
-					PsGetModuleValidationCache=$PsGetModuleValidationCache
+				#Validate All Modules
+				$AllModulesPath = New-Object System.Collections.ArrayList -ErrorAction Stop
+				foreach ($ModulePath in $SolutionConfig.SolutionStructure.ModulesPath)
+				{
+					$null = $AllModulesPath.AddRange((Get-ChildItem -Path $ModulePath.SourcePath -Directory -ErrorAction Stop))
 				}
-				Build-PSModule @BuildPSSolutionModule_Params -ErrorAction Stop
+				priv_Validate-Module -SourcePath $AllModulesPath -ModuleValidationCache $ModuleValidationCache
+
+				foreach ($ModulePath in $SolutionConfig.SolutionStructure.ModulesPath)
+				{
+					$Modules = Get-ChildItem -Path $ModulePath.SourcePath -Directory -ErrorAction Stop
+					$BuildPSSolutionModule_Params = @{
+						SourcePath=$Modules
+						DestinationPath=$ModulePath.BuildPath
+						ResolveDependancies=$SolutionConfig.Build.AutoResolveDependantModules
+						PSGetRepository=$SolutionConfig.Packaging.PSGetSearchRepositories
+						CheckCommandReferences=$SolutionConfig.Build.CheckCommandReferences.Enabled
+						CheckCommandReferencesConfiguration=$SolutionConfig.Build.CheckCommandReferences
+						ModuleValidationCache=$ModuleValidationCache
+						UpdateModuleReferences=$SolutionConfig.Build.UpdateModuleReferences
+						PsGetModuleValidationCache=$PsGetModuleValidationCache
+					}
+					Build-PSModule @BuildPSSolutionModule_Params -ErrorAction Stop
+				}
+		
+				Write-Verbose "Build PSModules completed"
 			}
-      
-			Write-Verbose "Build PSModules completed"
+			else
+			{
+				Write-Verbose "Build PSModules skipped"
+			}
 		}
 		catch
 		{
